return more info about shoppinglistitem

+++ RestRecipeApp.Persistence/Repositories/ShoppingListRepository.cs
@@ -1,5 +1,4 @@
 using Microsoft.EntityFrameworkCore;
-using RestRecipeApp.Core.RequestDto;
 using RestRecipeApp.Persistence.Models;
 
 namespace RestRecipeApp.Persistence.Repositories;
@@ -15,26 +14,46 @@ public class ShoppingListRepository : IShoppingListRepository
         _ingredientRepository = ingredientRepository;
     }
 
-    public async Task<ShoppingList?> GetShoppingListById(int id)
+    public async Task<GetShoppingList?> GetShoppingListById(int id)
     {
-        // Todo return ingredient object in items
-        var shoppingList =  await _context.ShoppingLists
+        var shoppingList = await _context.ShoppingLists
+            .Include(sl => sl.Items)
             .FirstOrDefaultAsync(sl => sl.ShoppingListId == id);
+
         if (shoppingList == null)
         {
             return null;
         }
-        var items = await _context.ShoppingListItems.FromSqlRaw()
+
+        var items = GetShoppingListItems(shoppingList);
+
+        return new GetShoppingList
+        {
+            ShoppingListId = shoppingList.ShoppingListId,
+            RecipeId = shoppingList.RecipeId,
+            Items = items
+        };
     }
 
-    public async Task<List<ShoppingList>> GetAll()
+
+    public async Task<List<GetShoppingList>> GetAll()
     {
-        return await _context.ShoppingLists
+        var shoppingLists = await _context.ShoppingLists
             .Include(sl => sl.Items)
             .ToListAsync();
+        return shoppingLists.Select(x =>
+        {
+            var items = GetShoppingListItems(x);
+            return new GetShoppingList
+            {
+                ShoppingListId = x.ShoppingListId,
+                RecipeId = x.RecipeId,
+                Items = items
+            };
+        }).ToList();
     }
 
-    public async Task<ShoppingList> CreateShoppingList(int recipeId)
+    public async Task<GetShoppingList?> CreateShoppingList(int recipeId)
     {
         var ingredients = await _ingredientRepository.GetIngredients(recipeId);
         var items = ingredients.Select(x => new ShoppingListItem
@@ -48,14 +67,44 @@ public class ShoppingListRepository : IShoppingListRepository
             RecipeId = recipeId,
             Items = items
         });
-        
+
         await _context.SaveChangesAsync();
-        return createdShoppingList.Entity;
+        var result = await GetShoppingListById(createdShoppingList.Entity.ShoppingListId);
+        return result;
     }
 
-    public async Task RemoveShoppingList(ShoppingList shoppingList)
+    public async Task RemoveShoppingList(int shoppingListId)
     {
+        var shoppingList = _context.ShoppingLists.FirstOrDefault(x => x.ShoppingListId == shoppingListId);
+        ;
+        if (shoppingList == null)
+        {
+            return;
+        }
+
         _context.ShoppingLists.Remove(shoppingList);
         await _context.SaveChangesAsync();
+    }
+
+    private List<GetShoppingListItem> GetShoppingListItems(ShoppingList shoppingList)
+    {
+        var items = shoppingList.Items.Select(item =>
+        {
+            var ingredient = _context.Ingredients
+                .Include(i => i.Product)
+                .FirstOrDefault(i => i.IngredientId == item.IngredientId);
+
+            return new GetShoppingListItem
+            {
+                ShoppingListItemId = item.ShoppingListItemId,
+                ShoppingListId = item.ShoppingListId,
+                IngredientId = item.IngredientId,
+                Amount = ingredient.Amount,
+                UnitOfMeasurement = ingredient.UnitOfMeasurement,
+                IngredientName = ingredient.Product.Name,
+                IsBought = item.IsBought
+            };
+        }).ToList();
+        return items;
     }
 }
